[Feat] 숙소 상세 페이지 및 테스트 컨테이너 적용 (#10)
* clear git cache

* feat: 각 세부 정보 쿼리 분리해서 설계
* 숙소 + 주소
* 이미지
* 편의시설

* refactor: 이름 수정

* feat: 리뷰 쿼리 및 dto 설계
* 숙소 -> 예약 -> 리뷰 순으로 참조

* refactor: 이미지 id도 가져오도록 수정

* feat: 응답 Dto 추가, respository 구조 통일

* feat: AccommodationService 구현

* refactor: 코드 구조 리팩토링

* feat: 숙소 상세 조회 컨트롤러 구현

* refactor: 쿼리용 DTO를 record로 분리

* test: 테스트 컨테이너 적용, 숙소 상세 조회 테스트 작성

+++ backend/src/main/java/com/team3/airdnd/payment/domain/Payment.java
@@ -1,6 +1,6 @@
 package com.team3.airdnd.payment.domain;
 
-import com.team3.airdnd.accomodation.domain.Reservation;
+import com.team3.airdnd.accommodation.domain.Reservation;
 import jakarta.persistence.*;
 import lombok.*;
 
@@ -12,6 +12,7 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 @NoArgsConstructor
 public class Payment {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
@@ -24,7 +25,7 @@ public class Payment {
     @JoinColumn(name = "payment_method_id", nullable = false)
     private PaymentMethod paymentMethod;
 
-    private Integer amount;
+    private Long amount;
 
     @Column(name = "paid_at", nullable = false)
     private LocalDateTime paidAt;
